+++ include/utils.hpp
@@ -20,15 +20,15 @@ namespace fastllama {
         }
         
         constexpr auto operator""_GiB (long double  size) noexcept {
-            return size * (1024.0 * 1024.0 * 1024.0);
+            return size * static_cast<long double>(1024.0 * 1024.0 * 1024.0);
         }
 
         constexpr auto operator""_MiB (long double  size) noexcept {
-            return size * (1024.0 * 1024.0);
+            return size * static_cast<long double>(1024.0 * 1024.0);
         }
         
         constexpr auto operator""_KiB (long double  size) noexcept {
-            return size * 1024.0;
+            return size * static_cast<long double>(1024.0);
         }
 
     } // namespace literals
